Rename utils.ts to usePaginatedEventsByIdsQuery.tsx
